pyproject.toml
[project]
name = "ynab-obsidian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.14.0",
    "ynab>=1.9.0",
]
